# Improve performance of SGEMM on Arm Cortex-A53